chore(tests): Remove old e2e tests and rename e2e2 to e2e
