+++ hack/dev-env/reverse-tunnel/README.md
@@ -20,7 +20,7 @@ To enable the configuration:
 
 ```
 # Setup environment as usual
-make setup-e2e2
+make setup-e2e
 
 # Installs Rathole Deployment/Services on vcluster, update 'managed-agent' cluster Secret to point to Rathole Deployment, and start local rathole container
 ./hack/dev-env/reverse-tunnel/setup.sh
@@ -31,14 +31,14 @@ To test the configuration:
 
 ```
 # Start E2E tests
-make start-e2e2
+make start-e2e
 ```
 
 
 
 ### Configuration
 
-Running 'setup-e2e2', followed by 'reverse-tunnel/setup.sh', enables this configuration:
+Running 'setup-e2e', followed by 'reverse-tunnel/setup.sh', enables this configuration:
 
 
 'vcluster-control-plane' vcluster (on k8s cluster):
