+++ test/e2e/delete_test.go
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package e2e2
+package e2e
 
 import (
 	"testing"
 	"time"
 
-	"github.com/argoproj-labs/argocd-agent/test/e2e2/fixture"
+	"github.com/argoproj-labs/argocd-agent/test/e2e/fixture"
 
 	"github.com/argoproj/argo-cd/v3/pkg/apis/application/v1alpha1"
 	argoapp "github.com/argoproj/argo-cd/v3/pkg/apis/application/v1alpha1"
