+++ test/e2e/fixture/argoclient.go
